+++ vue/src/views/experiment/ExperimentView.vue
@@ -15,6 +15,26 @@
  **/
 import TabsHeader from './components/TabsHeader.vue'
 import ExperimentLayout from '@swanlab-vue/layouts/ExperimentLayout.vue'
+import { computed, provide } from 'vue'
+import { useProjectStore } from '@swanlab-vue/store'
+import { useRoute } from 'vue-router'
+const route = useRoute()
+const projectStore = useProjectStore()
+
+// ---------------------------------- 获取当前实验的配置 ----------------------------------
+const experiment = computed(() => {
+  return projectStore.experiments?.find((item) => item.experiment_id === experimentId.value)
+})
+
+const experimentColor = computed(() => {
+  return experiment.value.color
+})
+
+const experimentId = computed(() => Number(route.params.experimentId))
+
+provide('experiment', experiment)
+provide('experimentId', experimentId)
+provide('experimentColor', experimentColor)
 </script>
 
 <style lang="scss" scoped></style>
