+++ vue/src/views/experiment/components/TabsHeader.vue
@@ -18,7 +18,7 @@
  * @file: TabsHeader.vue
  * @since: 2023-12-09 21:44:40
  **/
-const tabs = ['index', 'chart', 'log']
+const tabs = ['index', 'chart']
 </script>
 
 <style lang="scss" scoped>
