+++ vue/src/views/experiment/pages/IndexPage.vue
@@ -1,4 +1,6 @@
-<template>111</template>
+<template>
+  {{ experiment }}
+</template>
 
 <script setup>
 /**
@@ -6,6 +8,9 @@
  * @file: ExperimentIndex.vue
  * @since: 2023-12-09 20:37:45
  **/
+import { inject } from 'vue'
+
+const experiment = inject('experiment')
 </script>
 
 <style lang="scss" scoped></style>
